tests/wpt/tests/dom/nodes/getElementsByClassName-04.htm

<!doctype html>
<html class="a">
 <head>
  <title>document.getElementsByClassName(): changing classes</title>
  <script src="/resources/testharness.js"></script>
  <script src="/resources/testharnessreport.js"></script>
 </head>
 <body class="a">
  <div id="log"></div>
  <script>
   test(function() {
          var collection = document.getElementsByClassName("a");
          document.body.className += "\tb";
          assert_array_equals(collection, [document.documentElement, document.body]);
        })
  </script>
 </body>
</html>
